data base/extracao_de_dados/requirements.txt
requests
pymongo
python-dotenv